I/O

PeriodicGraphEmbeddings.export_vtf — Function

function export_vtf(file::AbstractString, pge::PeriodicGraphEmbedding3D{T}, types=nothing, repeatedges=6, colorname=false, tostring=string, atomnumof==(a,i)->(a isa Integer ? a : i)) where T

Export a PeriodicGraphEmbedding3D to a .vtf file (readable by VMD).

If specified, types is a list of types for each vertex of pge. Each type is converted to string by the tostring function. The atomnumof function takes two arguments ty and i where ty is a type and i is the number of the vertex, and return an Int representing an atom number.

PeriodicGraphEmbeddings.export_cgd — Function

export_cgd(file, pge::PeriodicGraphEmbedding, name=basename(splitext(file)[1]), append=false)
export_cgd(file, g::PeriodicGraph, name=basename(splitext(file)[1]), append=false)

Export a PeriodicGraph or a PeriodicGraphEmbedding to a .cgd file (readable by Systre).

If append is set, the graph is added at the end of the file.